Extract collection choosing switch statement.
