+++ lib/engines/classic/classic-file-info.js
@@ -1,12 +1,7 @@
 var FileInfo = require('../../models/file-info');
 
-var PLURAL_OVERRIDES = {
-  'mirage': 'mirage',
-  'html': 'html'
-};
 var ClassicFileInfo = FileInfo.extend({
-  fileInfoType: 'ClassicFileInfo',
-  plural_overrides: PLURAL_OVERRIDES
+  fileInfoType: 'ClassicFileInfo'
 });
 
 module.exports = ClassicFileInfo;
