+++ lib/engines/classic/test-file-info.js
@@ -1,5 +1,6 @@
 var path = require('path');
 var ClassicFileInfo = require('./classic-file-info');
+var calculateCollectionInfo = require('../../utils/calculate-collection-info');
 
 var TestFileInfo = ClassicFileInfo.extend({
   populate: function() {
@@ -36,7 +37,10 @@ var TestFileInfo = ClassicFileInfo.extend({
   },
 
   populateCollection: function() {
-    this._super.populateCollection.apply(this, [this.testSubjectType]);
+    var values = calculateCollectionInfo(this.testSubjectType);
+
+    this.collection = values.collection;
+    this.collectionGroup = values.collectionGroup;
   }
 });
