+++ lib/models/file-info.js
@@ -1,6 +1,7 @@
 var path = require('path');
 var inflection = require('inflection');
 var CoreObject = require('core-object');
+var calculateCollectionInfo = require('../utils/calculate-collection-info');
 
 var FileInfo = CoreObject.extend({
   type: 'FileInfo',
@@ -73,71 +74,10 @@ var FileInfo = CoreObject.extend({
 
   populateCollection: function(_type) {
     var type = _type || this.type;
-    var pluralizedType = inflection.pluralize(type, this.plural_overrides[type]);
-    var collection, collectionGroup;
+    var values = calculateCollectionInfo(type);
 
-    switch (type) {
-    case 'service':
-      collection = 'services';
-      collectionGroup = '';
-      break;
-
-    case 'util':
-    case 'mixin':
-      collection = 'utils';
-      collectionGroup = '';
-      break;
-
-    case 'authenticator':
-    case 'authorizer':
-    case 'session-store':
-      collection = pluralizedType;
-      collectionGroup = 'simple-auth';
-      break;
-
-    case 'adapter':
-    case 'serializer':
-    case 'model':
-      collection = 'models';
-      collectionGroup = 'data';
-      break;
-
-    case 'transform':
-      collection = 'transforms';
-      collectionGroup = 'data';
-      break;
-
-    case 'route':
-    case 'controller':
-    case 'template':
-      collection = 'routes';
-      collectionGroup = 'ui';
-      break;
-
-    case 'helper':
-    case 'component':
-      collection = 'elements';
-      collectionGroup = 'ui';
-      break;
-
-    case 'style':
-      collection = 'styles';
-      collectionGroup = 'ui';
-      break;
-
-    case 'initializer':
-    case 'instance-initializer':
-      collection = pluralizedType;
-      collectionGroup = 'init';
-      break;
-
-    default:
-      collection = pluralizedType;
-      collectionGroup = '';
-    }
-
-    this.collection = collection;
-    this.collectionGroup = collectionGroup;
+    this.collection = values.collection;
+    this.collectionGroup = values.collectionGroup;
   }
 });
